adding indexing / index advisor tutorial to user profile tutorial
