+++ modules/profile-store/examples/dotnet/UserProfileExample/UserProfileExample.Models/EventType.cs
@@ -1,0 +1,9 @@
+﻿namespace UserProfileExample.Models
+{
+    public enum EventType
+    {
+        ProductViewed = 0,
+        ProductAddedToCart = 1,
+        ProfileUpdated = 2
+    }
+}
