+++ modules/profile-store/examples/dotnet/UserProfileExample/UserProfileExample.Models/User.cs
