+++ modules/profile-store/examples/dotnet/UserProfileExample/UserProfileExample.Models/UserProfileExample.Models.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Faker.Net" Version="1.1.1" />
+  </ItemGroup>
+
+</Project>
