+++ modules/profile-store/examples/dotnet/UserProfileExample/UserProfileExample.SampleGenerator/UserProfileExample.SampleGenerator.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\UserProfileExample.Models\UserProfileExample.Models.csproj" />
+  </ItemGroup>
+
+</Project>
