+++ modules/profile-store/examples/dotnet/UserProfileExample/UserProfileExample/UserProfileExample.csproj
@@ -16,6 +16,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\UserProfileExample.Models\UserProfileExample.Models.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
     <Content Update="nlog.config">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
