+++ modules/profile-store/pages/dotnet.adoc
@@ -134,7 +134,7 @@ Then, let's create a simple entity called User:
 
 [source,C#,indent=0]
 ----
-include::../examples/dotnet/UserProfileExample/UserProfileExample/Models/User.cs[tag=User]
+include::../examples/dotnet/UserProfileExample/UserProfileExample.Models/User.cs[tag=User]
 ----
 
 With that class in mind, we can also create a very simple repository class. Right now, we only need two methods. One method to save a new User and one method to get a user by ID:
@@ -208,7 +208,7 @@ Let's first expand our example to something closer to the real world by adding a
 
 [source,C#,indent=0]
 ----
-include::../examples/dotnet/UserProfileExample/UserProfileExample/Models/User.cs[tag=User2]
+include::../examples/dotnet/UserProfileExample/UserProfileExample.Models/User.cs[tag=User2]
 ----
 
 Even though we added some new attributes, our code has barely changed:
@@ -290,6 +290,11 @@ For `FindActiveUsersByFirstName` it's a little trickier, because the N1QL is gen
 ----
 include::../examples/dotnet/UserProfileExample/UserProfileExample/Models/indexes.n1ql[tag=ix_findActiveUsersByFirstName]
 ----
+
+Creating the right indexes can be difficult. If you are having trouble creating the right index, here are two paths you can take to get help:
+
+* Starting in Couchbase Server 6.5, there is an index advisor that can automatically make index recommendations for you. To learn more about this, read the link:link:http://docs.couchbase.com/tutorials/profile-store/indexing.html[User Profile Indexing tutorial]. This tutorial uses the same type of data that you've already been creating if you've been following along with this User Profile Tutorial.
+* Check out the link:https://forums.couchbase.com/c/n1ql[N1QL category on the Couchbase Forums]. You can post a question about N1QL queries or indexes, and the community there can often help you.
 
 Now that you know how to query your user profile store, let's revisit our list of non-functional requirements:
